Write an imporved HDMAEN latch test
This new test tries to write to HDMAEN 2 master cycles after the
previous test.

I tried to remove the test dependence on H-Counter IRQ but I was unable
to get the inner loop to execute in exactly 2 scanlines of time
on bsnes-plus/mesen-s/bsnes (Maybe the DRAM refresh is shifting it 2/4
cycles every 5-8 tests? IDK)

So I restarted from scratch. This time I used the H-Counter IRQ, wai and
hand crafted loops.  After tinkering with the timings a bit (in
bsnes-plus) I have gotten the `sta HDMAEN` to be executed at a more
predictable manner.