Stabilize the horizontal position of the IRQ ISR
This commit stabilizes the IRQ ISRs by creating a new NMI ISR that is 4
master cycles slower on non-interlace frames.

`hdmaen_latch_test_2` was not updated because adding the new NmiHandler
made the flickering worse.

Testing with Mesen-S's event viewer confirms the IRQ ISR is executed at
the same dot position on each frame in the INIDISP tests.

This commit has also been tested on a real console and I've confirmed
the IRQs are stable on the updated tests (Super Famicom, CPUv2, PPUv1/3)